feat-103: rename some files for consistency
